backend/requirements.txt
# LangChain + LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1

# Loaders (PDF, DOCX, HTML)
unstructured>=0.10.0
pypdf>=3.0.0
python-docx>=0.8.0
docx2txt>=0.8
pymupdf>=1.23.0
beautifulsoup4>=4.12.0
html5lib>=1.1

# Embeddings & Transformers
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# ChromaDB for VectorStore
chromadb>=0.4.0

# Data Validation
pydantic>=2.0.0

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
gunicorn>=21.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
black>=23.0.0
flake8>=6.1.0

# HuggingFace
huggingface-hub>=0.17.0

# Wikipedia
wikipedia>=1.4.0